requirements.txt
fastapi==0.78.0
uvicorn[standard]==0.17.6
openai==0.28.1
pydantic==1.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
